# apps/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/onboarding/step/{stepKey}": {
      "post": {
        "operationId": "OnboardingController_submitStep",
        "parameters": [
          {
            "name": "stepKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "onboarding"
        ]
      }
    },
    "/onboarding/skip/{stepKey}": {
      "post": {
        "operationId": "OnboardingController_skip",
        "parameters": [
          {
            "name": "stepKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "onboarding"
        ]
      }
    },
    "/onboarding/placement/{trackId}/start": {
      "post": {
        "operationId": "OnboardingController_placementStart",
        "parameters": [
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "onboarding"
        ]
      }
    },
    "/onboarding/placement/{trackId}/submit": {
      "post": {
        "operationId": "OnboardingController_placementSubmit",
        "parameters": [
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "onboarding"
        ]
      }
    },
...
    "/users/username-available": {
      "get": {
        "operationId": "UsersController_usernameAvailable",
        "parameters": [
          {
            "name": "u",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/u/{username}": {
      "get": {
        "operationId": "PublicProfileController_publicProfile",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "profiles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Learnix API",
    "description": "Accounts Centre + Personalized Onboarding — global audience.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lx_access"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}